refactor.auth-user
